backend/requirements.txt
flask~=3.0.3
langchain-google-genai
langchain-community
langchain-text-splitters
langchain-community-embeddings
langchain-community-document-loaders
langchain~=0.2.1
pdfplumber
chromadb
python-dotenv~=1.0.1
